refactor(domain): #341 — relocate parsed POD config types to crap-core::domain::config
Move the parsed-projection POD types — `FileConfig`, `OutputConfig`,
`LangConfig`, `ViewPreset` — out of `adapters/config.rs` into a new
`crap-core/src/domain/config.rs`. These are the language-agnostic shape
the core merges and analyzes: their fields are domain types
(`ThresholdPreset`, `ComplexityMetric`, `PathBuf`, …), not wire strings,
so they belong in `domain/` alongside the other serde-bearing domain
types. Every adapter (Rust, TypeScript) produces the same `FileConfig`.

`adapters::config` re-exports the four types via `pub use`, so the two
in-crate consumers (`cli::mod`, `cli::view_args`) and `parse_config`
(which constructs them) compile with zero source changes.

Not moved: `ThresholdOverride` (already in `domain/threshold.rs`),
`SrcSpec` (untagged JsonSchema wire enum), `ConfigDiscovery` (loader
artifact), and the `ConfigSchema` wire/schema family (schemars +
documented) — all stay in the adapters layer.

`FileConfig`'s rustdoc no longer intra-doc-links the adapter-layer
`ConfigSchema` type (a `domain → adapters` link would be a layering
inversion and would strand under `RUSTDOCFLAGS=-D warnings`); the prose
states the distinction directly. The sibling `[`ViewPreset`]` /
`[`LangConfig`]` links move with the types and still resolve.

Pure organizational move — no behavior change.
